+++ app/admin/__init__.py
@@ -1,3 +1,3 @@
-from .city_admin import EmployeeAdmin, ClientAdmin
+from .user_admin import EmployeeAdmin, ClientAdmin
 
 admin_classes = (EmployeeAdmin, ClientAdmin,)
